remove the old Main component
