+++ src/App.js
@@ -1,7 +1,6 @@
 import './App.css';
 import Header from "./components/Header.jsx";
-//import Main from "./components/Main.jsx";
-import MainNew from "./components/MainNew.jsx";
+import Main from "./components/Main.jsx";
 import Footer from "./components/Footer.jsx";
 
 function App() {
